requirements.txt
beautifulsoup4==4.11.1
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
filelock==3.7.1
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
Pillow==9.2.0
PySocks==1.7.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2